extracted form sections: car details and driver details from settings view to separate views.
