+++ CarCompanion/SettingsView.swift
@@ -17,62 +17,15 @@ struct SettingsView: View {
     @State var owner: String
     @State var licensePlate: String
     @State var driverLicense: String
-    
-    var latestYear = { () -> Int in
-        let date = Date()
-        let calender = Calendar.current
-        let component = calender.dateComponents([.year], from: date)
-        return component.year!
-        
-    }()
-    
+            
     var body: some View {
         NavigationView {
             VStack {
                 Spacer()
                 Form {
-                    Group {
-                       
-                        TitleView(title: "CAR DETAILS")
-                            .padding(.top, 20)
-                            .listRowBackground(Color.clear)
-                            .listRowInsets(EdgeInsets())
-                        Section("Make") {
-                            TextField("What brand is this car?", text: $make)
-                        }
-                        Section("Model") {
-                            TextField("What model is this car?", text: $model)
-                        }
-                        Section("Year") {
-                            Picker("What year is your car make?", selection: $year) {
-                                ForEach((1900...latestYear).reversed(), id: \.self) {
-                                    Text(String($0))
-                                }
-                            }
-                        }
-                        Section("License plate") {
-                            TextField("What is the Car License Plate Number", text: $licensePlate)
-                        }
-                        Section("ODOMETER (in KM)") {
-                            TextField("What is your car odometer reading status?", value: $odometer, format: .number)
-                        }
-                        Section("Fuel Capacity (in Litre") {
-                            TextField("What is the car fuel capacity?", value:$fuelCapacity, format: .number)
-                        }
-                    }
+                    CarDetailsView(make: $make, model: $model, year: $year, odometer: $odometer, fuelCapacity: $fuelCapacity, licensePlate: $licensePlate)
                     
-                    Group {
-                        TitleView(title: "DRIVER DETAILS")
-                            .padding(.top, 15)
-                            .listRowBackground(Color.clear)
-                            .listRowInsets(EdgeInsets())
-                        Section("Owner's Full Name") {
-                            TextField("Who is the car Owner?", text: $owner)
-                        }
-                        Section("Driver's License") {
-                            TextField("Enter Your Driver's Lincense Number", text: $driverLicense)
-                        }
-                    }
+                    DriverDetailsView(owner: $owner, driverLicense: $driverLicense)
                     VStack {
                         AppButton(text: "Save", color: .blue, action: {print("save")}, width: 300, height: 40)
                     }
@@ -90,3 +43,5 @@ struct SettingsView_Previews: PreviewProvider {
         SettingsView(make: "", model: "", year: 2020, odometer: 0.0, fuelCapacity: 0, owner: "", licensePlate: "", driverLicense: "")
     }
 }
+
+
